+++ src/backend/joanie/tests/core/test_models_order.py
@@ -1270,12 +1270,20 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
         self,
     ):
         """
-        We want to test when we submit a contract to the signature provider and
-        overtime the context has changed. Then, the method `submit_for_signature` should
-        trigger the method `delete_signing_procedure` to update the contract at the signature
-        provider.
-        In this specific case, we want to test what happens to the contract if it reaches a
-        TimeOut Error on the `delete_signing_procedure` method.
+        We test the behavior of `submit_for_signature` when a `ReadTimeout` error is caught
+        during the `delete_signing_procedure` API call. We simulate the case where the signature
+        provider takes a long time to process the deletion of a signature workflow on their side.
+        For this test, we have prepared, using responses, all the requests that will occur during
+        the two calls to `submit_for_signature` for an order.
+
+        When the condition for `should_be_resubmitted` is met, we simulate a `ReadTimeout`
+        during the `delete_signing_procedure` API call. The contract should then be reset before
+        submitting the new document for signature.
+
+        The first reference will have the value `wfl_id_fake_1`, and the second will have the value
+        `wfl_id_fake_2`. At the end of this test, our contract should have the value
+        `wfl_id_fake_2`, with the updated hash value. Additionally, we should see the title
+        change in the contract's definition (which triggered the new contract submission).
         """
         user = factories.UserFactory(
             email="johnnydo@example.fr",
@@ -1295,15 +1303,12 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
         factories.UserOrganizationAccessFactory.create_batch(
             3, organization=order.organization, role="owner"
         )
-        workflow_id = "wfl_id_fake"
-
-        ## Create workflow
-        create_workflow_api_url = (
-            "https://lex_persona.test01.com/api/users/usr_id_fake/workflows"
-        )
+        workflow_id = "wfl_id_fake_1"
+        hash_1 = "wpTD3tstfdt9XfuFK+sv4/y6fv3lx3hwZ2gjQ2DBrxs="
+        # Create workflow for the first document to sign
         responses.add(
             responses.POST,
-            create_workflow_api_url,
+            "https://lex_persona.test01.com/api/users/usr_id_fake/workflows",
             status=HTTPStatus.OK,
             json={
                 "created": 1696238245608,
@@ -1328,7 +1333,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                         "allowComments": True,
                         "hideAttachments": False,
                         "hideWorkflowRecipients": True,
-                        "id": "stp_J5gCgaRRY4NHtbGs474WjMkA",
+                        "id": "stp_id_fake",
                         "invitePeriod": None,
                         "isFinished": False,
                         "isStarted": False,
@@ -1366,8 +1371,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                 ),
             ],
         )
-
-        ## upload file to workflow
+        # Upload the file to workflow for the first document to sign on the contract
         upload_file_api_url = (
             f"https://lex_persona.test01.com/api/workflows/{workflow_id}"
             "/parts?createDocuments=true&ignoreAttachments=false"
@@ -1387,7 +1391,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                             {
                                 "contentType": "application/pdf",
                                 "filename": "contract_definition.pdf",
-                                "hash": "wpTD3tstfdt9XfuFK+sv4/y6fv3lx3hwZ2gjQ2DBrxs=",
+                                "hash": hash_1,
                                 "size": 123616,
                             }
                         ],
@@ -1397,7 +1401,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                         "userId": "usr_id_fake",
                         "viewAuthorizedGroups": ["grp_id_fake"],
                         "viewAuthorizedUsers": [],
-                        "workflowId": "wfl_id_fake",
+                        "workflowId": "wfl_id_fake_1",
                         "workflowName": "Heavy Duty Wool Watch",
                     }
                 ],
@@ -1406,7 +1410,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                     {
                         "contentType": "application/pdf",
                         "filename": "contract_definition.pdf",
-                        "hash": "wpTD3tstfdt9XfuFK+sv4/y6fv3lx3hwZ2gjQ2DBrxs=",
+                        "hash": hash_1,
                         "size": 123616,
                     }
                 ],
@@ -1416,11 +1420,10 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                     {
                         "Authorization": "Bearer token_id_fake",
                     },
-                )
+                ),
             ],
         )
-
-        ## Start procedure
+        ## Start signing procedure of the workflow
         start_procedure_api_url = (
             f"https://lex_persona.test01.com/api/workflows/{workflow_id}"
         )
@@ -1443,7 +1446,6 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                 ),
             ],
         )
-
         # Sign specific contract
         responses.add(
             responses.POST,
@@ -1460,7 +1462,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                     {
                         "allowComments": True,
                         "stepId": "stp_id_fake",
-                        "workflowId": "wfl_id_fake",
+                        "workflowId": workflow_id,
                     }
                 ],
                 "tenantId": "ten_id_fake",
@@ -1475,16 +1477,15 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                 ),
                 responses.matchers.json_params_matcher(
                     {
-                        "workflows": ["wfl_id_fake"],
-                    }
+                        "workflows": [workflow_id],
+                    },
                 ),
             ],
         )
-
-        # Get the invitation link
+        # Get the invitation link for the first document to sign
         responses.add(
             responses.POST,
-            "https://lex_persona.test01.com/api/workflows/wfl_id_fake/invite",
+            "https://lex_persona.test01.com/api/workflows/wfl_id_fake_1/invite",
             json={"inviteUrl": "https://example.com/invite?token=jwt_token"},
             status=HTTPStatus.OK,
             match=[
@@ -1498,66 +1499,240 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
                 ),
             ],
         )
+
+        # Get the invitation signature for the first document to sign
         invitation_url = order.submit_for_signature(user=user)
+
         self.assertEqual(
             invitation_url,
             "https://lex_persona.test01.com/?requestToken=eyJhbGciOiJIUzI1NiJ9"
             "#requestId=req_8KVKj7qNKNDgsN7Txx1sdvaT",
         )
+        self.assertEqual(order.contract.definition.title, "Contract grade 1")
+        self.assertEqual(order.contract.signature_backend_reference, "wfl_id_fake_1")
+        self.assertEqual(order.contract.definition_checksum, hash_1)
 
-        # Save the timestamp of the updated_on of the contract
-        first_attempt_submission = order.contract.updated_on
-        # Now we change the contract definition to trigger the should_be_resubmitted condition
-        # and call the method delete_signing_procedure
+        # Save the timestamp of the `updated_on` of the contract
+        contract_last_update_on_1 = order.contract.updated_on
+        # Change the contract definition title to trigger the `should_be_resubmitted` condition
         order.product.contract_definition.title = "You know nothing John Snow."
         order.product.contract_definition.save()
-
-        delete_api_endpoint_url = (
-            "https://lex_persona.test01.com/api/workflows/wfl_id_fake"
-        )
-
+        # Prepare the ReadTimeout on the `delete_signing_procedure` method
         responses.add(
             responses.DELETE,
-            delete_api_endpoint_url,
+            f"https://lex_persona.test01.com/api/workflows/{workflow_id}",
             body=ReadTimeout(
                 "Timeout occurred when trying to delete the signing procedure"
             ),
         )
-
-        # Now call submit_for_signature, which should trigger the timeout in
-        # `delete_signing_procedure`.
-        # The first_attempt_submission timestamp should be different than the
-        # second_attempt_submission because it's supposed to trigger the creation of a new
-        # workflow. And here it's not the case.
-        # They are the same, which tells us that no new contract gets submitted. Meaning that
-        # we never entered the condition where it's supposed to create a signature procedure
-        # and call the method `tag_submission_for_signature`
-        with mock.patch(
-            "joanie.core.models.contracts.Contract.tag_submission_for_signature"
-        ) as mock_tag_submission_for_signature:
-            with self.assertRaises(ReadTimeout) as context:
-                order.submit_for_signature(user=user)
-
-        self.assertEqual(
-            str(context.exception),
-            "Timeout occurred when trying to delete the signing procedure",
+        # Prepare the data for the new document to sign on the contract
+        new_workflow_id = "wfl_id_fake_2"
+        hash_2 = "wpTD3tstfdt9XfuFK+sv4/y6fv3lx3hwZ2gjQ2Dqsdxs="
+        responses.add(
+            responses.POST,
+            "https://lex_persona.test01.com/api/users/usr_id_fake/workflows",
+            status=HTTPStatus.OK,
+            json={
+                "created": 1696238245608,
+                "currentRecipientEmails": [],
+                "currentRecipientUsers": [],
+                "description": "Contract Definition",
+                "email": order.owner.email,
+                "firstName": order.owner.first_name,
+                "groupId": "grp_id_fake",
+                "id": new_workflow_id,
+                "lastName": ".",
+                "logs": [],
+                "name": "Contract Definition",
+                "notifiedEvents": [
+                    "recipientRefused",
+                    "recipientFinished",
+                    "workflowFinished",
+                ],
+                "progress": 0,
+                "steps": [
+                    {
+                        "allowComments": True,
+                        "hideAttachments": False,
+                        "hideWorkflowRecipients": True,
+                        "id": "stp_id_fake",
+                        "invitePeriod": None,
+                        "isFinished": False,
+                        "isStarted": False,
+                        "logs": [],
+                        "maxInvites": 0,
+                        "recipients": [
+                            {
+                                "consentPageId": "cop_id_fake",
+                                "country": order.main_invoice.recipient_address.country.code.upper(),  # pylint: disable=line-too-long
+                                "email": "johnnydoe@example.fr",
+                                "firstName": "Johnny",
+                                "lastName": ".",
+                                "preferredLocale": "fr",
+                            }
+                        ],
+                        "requiredRecipients": 1,
+                        "sendDownloadLink": True,
+                        "stepType": "signature",
+                        "validityPeriod": 1296000000,
+                    }
+                ],
+                "tenantId": "ten_id_fake",
+                "updated": 1696238245608,
+                "userId": "usr_id_fake",
+                "viewAuthorizedGroups": ["grp_id_fake"],
+                "viewAuthorizedUsers": [],
+                "watchers": [],
+                "workflowStatus": "stopped",
+            },
+            match=[
+                responses.matchers.header_matcher(
+                    {"Authorization": "Bearer token_id_fake"}
+                )
+            ],
         )
-        mock_tag_submission_for_signature.assert_not_called()
+        # Upload the document to sign of the contract
+        responses.add(
+            responses.POST,
+            f"https://lex_persona.test01.com/api/workflows/{new_workflow_id}/parts",
+            status=HTTPStatus.OK,
+            json={
+                "documents": [
+                    {
+                        "created": 1696238255558,
+                        "groupId": "grp_id_fake",
+                        "id": "doc_id_fake",
+                        "parts": [
+                            {
+                                "contentType": "application/pdf",
+                                "filename": "contract_definition.pdf",
+                                "hash": hash_2,
+                                "size": 123616,
+                            }
+                        ],
+                        "signatureProfileId": "sip_profile_id_fake",
+                        "tenantId": "ten_id_fake",
+                        "updated": 1696238255558,
+                        "userId": "usr_id_fake",
+                        "viewAuthorizedGroups": ["grp_id_fake"],
+                        "viewAuthorizedUsers": [],
+                        "workflowId": new_workflow_id,
+                        "workflowName": "Heavy Duty Wool Watch",
+                    }
+                ],
+                "ignoredAttachments": 0,
+                "parts": [
+                    {
+                        "contentType": "application/pdf",
+                        "filename": "contract_definition.pdf",
+                        "hash": hash_2,
+                        "size": 123616,
+                    }
+                ],
+            },
+            match=[
+                responses.matchers.header_matcher(
+                    {"Authorization": "Bearer token_id_fake"}
+                )
+            ],
+        )
+        start_procedure_response_data = get_expected_workflow_payload("started")
+        start_procedure_response_data["id"] = new_workflow_id
+        responses.add(
+            responses.PATCH,
+            f"https://lex_persona.test01.com/api/workflows/{new_workflow_id}",
+            status=HTTPStatus.OK,
+            json=start_procedure_response_data,
+            match=[
+                responses.matchers.header_matcher(
+                    {
+                        "Authorization": "Bearer token_id_fake",
+                    },
+                ),
+                responses.matchers.json_params_matcher(
+                    {
+                        "workflowStatus": "started",
+                    }
+                ),
+            ],
+        )
+        # Sign specific contract for the new document to sign
+        responses.add(
+            responses.POST,
+            "https://lex_persona.test01.com/api/requests/",
+            json={
+                "consentPageId": "cop_id_fake",
+                "consentPageUrl": (
+                    "https://lex_persona.test01.com/?"
+                    "requestToken=eyJhbGciOiJIUzI1NiJ9#requestId=req_8KVKj7qNKNDgsN7Txx1sdvaT"
+                ),
+                "created": 1696238302063,
+                "id": "req_id_fake",
+                "steps": [
+                    {
+                        "allowComments": True,
+                        "stepId": "stp_id_fake",
+                        "workflowId": new_workflow_id,
+                    }
+                ],
+                "tenantId": "ten_id_fake",
+                "updated": 1696238302063,
+            },
+            status=HTTPStatus.OK,
+            match=[
+                responses.matchers.header_matcher(
+                    {
+                        "Authorization": "Bearer jwt_token",
+                    },
+                ),
+                responses.matchers.json_params_matcher(
+                    {
+                        "workflows": [new_workflow_id],
+                    }
+                ),
+            ],
+        )
+        # Invite to sign url for the new document to sign
+        responses.add(
+            responses.POST,
+            "https://lex_persona.test01.com/api/workflows/wfl_id_fake_2/invite",
+            json={"inviteUrl": "https://example.com/invite?token=jwt_token"},
+            status=HTTPStatus.OK,
+            match=[
+                responses.matchers.header_matcher(
+                    {
+                        "Authorization": "Bearer token_id_fake",
+                    },
+                ),
+                responses.matchers.json_params_matcher(
+                    {"recipientEmail": "johnnydo@example.fr"}
+                ),
+            ],
+        )
 
-        second_attempt_submission = order.contract.updated_on
+        # Get the invitation signature for the new document to sign
+        with self.assertLogs("joanie") as logger:
+            order.submit_for_signature(user=user)
 
-        self.assertEqual(first_attempt_submission, second_attempt_submission)
+        # We should find the in the logger message the reference wfl_id_fake_1 being deleted
+        self.assertLogsEquals(
+            logger.records,
+            [
+                (
+                    "ERROR",
+                    "Signature Provider is taking a while on deletion of reference wfl_id_fake_1.",
+                    {"order": dict, "product": dict},
+                )
+            ],
+        )
 
+        # Check we have the latest data from db for the contract
         contract = order.contract
+        contract.refresh_from_db()
+        contract_last_update_on_2 = contract.updated_on
 
+        self.assertNotEqual(contract_last_update_on_1, contract_last_update_on_2)
         self.assertIsNotNone(contract.submitted_for_signature_on)
-        self.assertIsNotNone(contract.context)
-        self.assertIsNotNone(contract.definition_checksum)
-        self.assertIsNotNone(contract.signature_backend_reference)
-
-        contract.reset_submission_for_signature()
-
-        self.assertIsNone(contract.submitted_for_signature_on)
-        self.assertIsNone(contract.context)
-        self.assertIsNone(contract.definition_checksum)
-        self.assertIsNone(contract.signature_backend_reference)
+        self.assertEqual(contract.signature_backend_reference, new_workflow_id)
+        self.assertEqual(contract.definition.title, "You know nothing John Snow.")
+        self.assertEqual(contract.definition_checksum, hash_2)